Move files and make success event write test pass
