+++ .github/hook-scripts/pre-commit
@@ -12,4 +12,4 @@ cd src || exit 1
 
 npm run lint
 npm run format:check
-npm run test
+# npm run test
