--- src/functions/services/events/eventWriter.ts
@@ -1,7 +1,0 @@
-import { ErrorOrSuccess } from "../../types/errorOrValue";
-
-export interface IWriteEvent {
-  writeEvent: (eventName: EventName) => ErrorOrSuccess<null>;
-}
-
-export type EventName = "DCMAW_ASYNC_CRI_5XXERROR";
